grrr.... fix compile error after latest fixes.

+++ src/target.cpp
@@ -109,7 +109,7 @@ namespace logog {
 
 		result = fwrite( &(*data), 1, data.size(), m_pFile );
 
-		if ( result != data.size() )
+		if ( (size_t)result != data.size() )
 			return -1;
 
 		return 0;
